expand serial rx buffer to 1024 bytes
This occupies half of all SRAM but since we aren't using it for anything
else really this seems fine. We need more than the default 64 bytes
because the machine on the other end of the serial port can easily
produce enough data at 115200 baud to overflow the 64 byte buffer and
cause us to lose data.